rename everything bus to truck
