+++ src/assets/truck/churner/churnerBackground.tsx
