+++ src/assets/truck/churner/churnerGlassReflection.tsx
