+++ src/assets/truck/churner/churnerLiquid.tsx
