+++ src/assets/truck/churner/churnerPipeGroup1.tsx
