+++ src/assets/truck/churner/churnerPipeGroup2.tsx
