+++ src/assets/truck/churner/churnerPipeGroup3.tsx
