+++ src/assets/truck/churner/churnerPropeller.tsx
