+++ src/assets/truck/churner/churnerPropellerMiddle.tsx
