+++ src/assets/truck/churner/churnerPropellerWings.tsx
