+++ src/assets/truck/churner/index.ts
