+++ src/assets/truck/create/createBackground.tsx
