+++ src/assets/truck/create/createBunny.tsx
