+++ src/assets/truck/create/createIceCreamCone.tsx
