+++ src/assets/truck/create/createIceCreamScoop.tsx
