+++ src/assets/truck/create/createPipeHorizontal.tsx
