+++ src/assets/truck/create/createPipeVertical.tsx
