+++ src/assets/truck/create/createSnow.tsx
