+++ src/assets/truck/create/index.ts
