+++ src/assets/truck/driver/driverFroggy.tsx
