+++ src/assets/truck/driver/index.ts
