+++ src/assets/truck/flair/flairBackground.tsx
