+++ src/assets/truck/flair/flairDuckyAndCherry.tsx
