+++ src/assets/truck/flair/index.ts
