+++ src/assets/truck/share/index.ts
