+++ src/assets/truck/share/shareBackground.tsx
