+++ src/assets/truck/truckBottom/index.ts
