+++ src/assets/truck/truckBottom/truckBottomPiece.tsx
