+++ src/assets/truck/truckBottom/truckBottomPieceLarge.tsx
