+++ src/assets/truck/truckBottom/truckBottomWheel.tsx
