+++ src/styles/truck/bottom.module.scss
