+++ src/styles/truck/share.module.scss
