+++ src/styles/truck/truck.module.scss
