+++ src/styles/truck/windowFour.module.scss
