+++ src/styles/truck/windowOne.module.scss
