+++ src/styles/truck/windowTwo.module.scss
